refactor: rename BlockThirdBody to BlockThirdBodyPointMass
