+++ .github/workflows/coverage.yml
@@ -74,7 +74,6 @@ jobs:
           xmake r testSpiceZpr
           xmake r testState
           xmake r testGUI
-          xmake r testHPOPEarth
           xmake test
           lcov --directory . --capture --include */src/* --output-file coverage.info
