--- include/AstCore/BlockThirdBody.hpp
@@ -1,1 +1,0 @@
-#include "../../src/AstCore/Propagator/HPOP/BlockAstro/BlockThirdBody.hpp"
