+++ include/AstCore/BlockThirdBodyPointMass.hpp
@@ -1,0 +1,1 @@
+#include "../../src/AstCore/Propagator/HPOP/BlockAstro/BlockThirdBodyPointMass.hpp"
