--- include/ast/BlockThirdBody.hpp
@@ -1,1 +1,0 @@
-#include "AstCore/BlockThirdBody.hpp"
