+++ include/ast/BlockThirdBodyPointMass.hpp
@@ -1,0 +1,1 @@
+#include "AstCore/BlockThirdBodyPointMass.hpp"
